fix plugin reload bug
Including vkhpp was exporting some symbols as UNIQUE, which makes the
library resident, i.e. it cannot be unloaded anymore for the lifetime of
the application.

To prevent this, we add a compiler flag, `--no-gnu-unique` to gcc when
compiling our plugins.

To check a library against unique symbols:

	readelf -sW pal_backend_vk/libpal_backend_vk.so | grep UNIQ

We should add a test which runs after each compile to check that there
are no unique symbols in our libs!