update renderpass creation
+ tweak resource state for each resource - the difficulty arises from
not being entirely sure what happens to a resource in terms of memory
access / shader stages when it is loaded/cleared when used for the very
first time in a renderpass.

we assume the driver internally issues memory barriers and needs to read
from a resource to load it.

+ we're also not 100% sure whether the layout transition occurs before
or after a resource is loaded.