add scratch allocator to encoder
+ adds scratch allocator to encoder - allocator is owned by current
frame in backend.

+ allocator is created when backend frames are created, and destroyed
when backend frames are destroyed, and reset when its backend frame is
reset (the latter is still todo)

+ data is recorded to allocator when frame is recorded, and will be used
for drawing when frame is being dispatched.

+ we will use the allocator to store uniform data, index data, and
vertex data. perhaps we should split this some more, but that's a
consideration for the future.

+ we may also consider using an allocator per command buffer (encoder),
as this will make it possible to have each encoder act as an isolate,
i.e. as a process which has no memory contention, as it owns all its
memory objects exclusively.

+ add some debug data to testapp