implement file watcher for shader source files
+ shader compiler builds a table for shader source files, and the
modules derived from each source. We take include files into account: if
an included file was changed, all modules which use it get a trigger to
recompile.

+ change the watch object for ApiRegistry - it appears that for our
library files (*.so) we can only reliably say that we need to reload
after we have detected a IN_WRITE_CLOSE event from the directory in
which the library file lives.

* change file_watcher to be more precise about the watched file -

    before this the file_watcher would watch the file's parent directory
    rather than the file directly. This would lead to the same watch
    being triggered for different files living in the same directory.