fix glsl source file watcher bug
fixes an issue where an included file, when changed, would only
trigger recompilation of the first module which depended on it, instead
of *all* modules which depended on it.

We're fixing this by, in the callback, calling the backend with the path
to the source file which has changed. The backend can then find out
which modules depend on this source file by looking up the
moduleDependencies list.