implement explicit vertex input setting
+ implements a user-selectable code path for explicitly setting vertex
input to shaders in a pipeline.

This makes sense in case you want to feed attributes from a single,
interleaved vertex buffer for example, or use more exotic buffer data
formats to feed your vertices.

If this path is not selected (not using this path is the default) then
vertex input for a pipeline is gathered automatically through reflection
of the vertex shader module.

When vertex input is automatically derived from shader reflection, there
is an 1:1 mapping from vertex input locations to attribute buffers, as
there is no way we can programmatically figure out whether vertex input
is to be read from interleaved or serialized buffer data.

NOTE: I'm not super happy with the detail implementation, the front-end
api is designed to be independent of vulkan, which is nice, but the
backend feels a bit too "clever".