implement imgui inputs
+ implements mouse and keyboard interactions with imgui

+ add callback wrappers from window to testapp so that when testapp
reloads, the function pointers for the callback (which go from GLFW to the window) don't need to be updated. Callback wrappers in window then de-reference the pointer which holds the address to the actual callback and call the callback.

these wrappers are ugly, and there must be a better way to make sure
that we can call callbacks at a fixed address, regardless of whether the
module which provides the callback implementation has been reloaded.