update filewatcher to respect filename
+ prior to this change file watcher would trigger if any file in a
directory had changed, and report a match for the first watcher which
was registered for a directory.

Because inotify reuses watch ids for multiple watchers to a directory,
checking only the inotify id is not correct.

we're now issueing watchers so that always the whole target directory is
watched, then we're checking up on the filename inotify reports as
matching, and we're using a match on the filename to trigger a watch
callback.

This allows us to keep all modules in the same directory.