implement auto-detect attachment format
automatically sets format for renderpass attachments to the format of the
referenced image if no explicit format for the renderpass was set.

When adding the backbuffer to the list of frame-available resources, its
format properties are copied too, which means this works for backbuffer,
too. No need to know the backbuffer image format in advance =)