add le::ResourceHandle facade
we add this facade to have a reliable way for LeResourceHandle to get
initialised to nullptr. Otherwise, we LeResourceHandles which were not
explicitly initialised will have garbage data, which might get
misinterpreted as a valid resource ID worst-case.

If we enforce to always use the c++ facade in the public interface we
have a way to enforce that handles must have been initialised before
use.