add check for invalid resourceHandle
if a LeResourceHandle evaluates to nullptr, this means that it has not
been declared with the renderer previously. As this is illegal
behaviour, we must print out an error message and panic.

This is not ideal, and I'm looking for a more systematic way to enforce
that ResourceHandles are unique and valid before use.