store swapchain dimensions with frame
+ since swapchain may change while frames are in flight, this means that
the width/height of the swapchain is a property of the frame. Therefore,
we acquire the swapchain dimensions when acquiring the swapchain image
and store it with the frame, so that any references to swapchain
dimensions use the same, frame-local value.

+ if renderpasses don't specify their dimensions explicitly, this means
that they shall have the same dimensions as the backbuffer. We implement
this by using the zero value for renderpass dimensions as a signal to
fill in the current frame's swapchain dimensions.