consolidate allocations to one translation unit
makes sure that there are no alloctions happening inside the header file
- these allocations would be placed in translation units which use this
header file as an interface, and the whole idea of a registry is that
this is a central location which consolidates all api allocations.

The Registry is also the only translation unit which is guaranteed not
to be reloaded.

We use a simple arena-allocator to store structs with function pointers
(that's how our apis are stored)