implement image loader first draft
- load pixels from image with help from stb_image

image loader is self-contained object which holds data for pixels. We
haven't yet implemented a path to load images which are other than 24 or
32 bit per pixel, and stb image has some nice things like float images
which could be interesting for the future, especially when it comes to
hdr images.

update testapp to use jpeg instead of headerified pixels for horse
image.