add minimal triangle draw app
+ minimal code necessary to draw a triangle

This app will serve as a canary in the ergonomic coalmine to tell us how
usable the general API is.

Ideally, it should be simple to extend this app, and it should also be
easy to discover features of the renderer from this app. If that is the
case, there are good reasons to assume that this will alos apply to
other applications built on the island API.