add assert against sparse bindings
+ bindings must not be sparse otherwise bad things will happen

This is not elegant, and we must find a way to recover from such shader
code. To implement protection against this, we need a higher level
concept, i.e. we need to be able to reject a combination of shader
modules which don't produce valid bindings. We need something like a
"linker" stage for the pipeline, and we must be able to report back that
pipeline generation failed without bailing out on the main concern.