implement staging allocator
+ staging allocator is frame-local, and allows us to make frame-local
allocations to accomodate uploading large amounts of source data for
bufferWrite or imageWrite. The staging allocator provides mapped
vkBuffers with CPU visible memory for staging during command recording.

These temporary buffers are used as the source buffers for copies to
GPU-local buffers/ or images, and discarded on frame clear, when the
copy operation can be guaranteed to have completed.