implement mipmap generation on imageWrite
+ imageWrite requires `resource_info_t` handle for the image resource,
as this describes the resource and using this is much more convenient
than to have separate parameters for each bit of the image. It also
ensures that the image is created/allocated with the correct settings
for mipmaps from the outset.

+ implements a cpu-side method to generate mipmap chains for images.
mipmap chains are generated for as many levels as the image requests in
its resource_info.