implement GPU generated mipmaps
+ on write_to_image, we auto-generate mipmap levels for images whose
resource_info has a miplevel count > 1.

+ These mipmaps are generated through a series of scaled blit
operations, which are enqueued in the backend. The backend also makes
sure that all image subresources of the image are in the correct layout
and properly synchonised after the operation.

+ note that because we use scaled blits to create mip levels, this means
that any image resources created with mip levels must also include the
USAGE_TRANSFER_SRC upon creation. We ensure this in the backend by
enforcing this flag when we see a request to create an image with mip
levels.