allocate image views for imgages used within frame
+ images used within a frame which have the flag 'STORAGE_IMAGE' set
will probably be referenced by the encoder, and as such need an image
view so that they can be accessed.

we therefore create a default image view for all images declared in a
frame so that all passes may use them to refer to images which are
available during that frame.

These image views are frame-local, which means their vulkan object has a
lifetime tied to the frame they live in, they get destroyed when the
frame comes around and gets recycled.