use resourceUsage instead of resource_info
changes the public api so that user needs to specify usage for each
resource used in a renderpass, and can specify resource info for
resources which are used across renderpasses and frames via
the renderModule interface.

The idea is to declare all persistent resources with their minimal
requirements using the renderModule interface, and then to specify
precise resource usage whenever using a resource with a renderpass.

This allows the backend to make better decisions about how to
synchronise resources.

The backend uses the infos for persistent resources consolidated with
usage information for each resource to allocate resources with all
needed attributes.