add le_mesh
implements a mesh type which is used to store structured mesh data. we
may implement methods to upload mesh data to the gpu in the future, but
more likely we will use this type as an internal store for mesh data.

i'm also refactoring the mesh_generator module. This module doesn't
store any data anymore, but is purely there to generate mesh data.