add unused code for forced viewport flipping
* we add some code for forced viewport flipping if we should ever decide
that we don't want to use the default vulkan coordinate system at all.

This is unlikely. But we're adding a link to an article by Sascha
Willems which describes the Vulkan coordinate system and why it might
make sense to have viewports with negative heights.

We have solved this problem in a different way, by allowing negative
viewport height to be set via the camera - this also enables the camera
to have consistent user experience because it can then detect flipping.