fix race condition
+ fixes an issue with a possible race condition around available fibers

We hand a fiber back to the pool of idle threads by setting its atomic
state to idle. Setting the thread state to idle means a change of
ownership. It must therefore be the last thing you do on this fiber. As
soon as the fiber becomes idle other threads might use this fiber, and
any settings you would then subsequently apply to the fiber will clash
with the new owner of the fiber.

Setting a fiber's state effectively means changing its ownership.

+ adds some documentation to inline assembly code

+ changes the sequencing for loading jobs - jobs are only popped from
the job queue *only* if there is a fiber available to actually execute
the job. Otherwise we would have ended with jobs popped but never
executed.