control access to pipeline manager internals
use shared_mutex to control access to graphics and compute pso cache -
since these caches might be accessed in the recording stage and also in
the processing stage there was a chance of a race on them.

we use `shared_mutex` because most access operations are reads, which
may happen concurrently, and there are only occasional writes, for which
these caches need to be locked exclusively.