make "textures" pass-local
when you define imageSamplers via `pass.sampleTexture`, the name defined
is local to pass scope, and is not shared with other renderpasses. This
means that temporary resources created from these names, such as
imageViews, and Samplers are local to the renderpass as well.

This change means that temporary resource names are limited in scope to
the renderpass for which they were defined in. This gives us more
freedom to choose names for resources, as the frame namespace does not
get polluted with resource names.