define swapchain image handle a global const
+ swapchain image resource handle is a symbolic handle, it will never
change, and can be defined at compile-time. It will always point to the
correct resource, since the backend resolves the resource to a physical
resource, internally.

By making the swapchain image handle available as a static constexpr,
`LE_SWAPCHAIN_IMAGE_HANDLE`, we make remove the need to query the
renderer for it all the time.