implement circle, ellipse, arc flattening
* we flatten circles and ellispes based on given tolerance. While this
works well for filled and outlined circles and filled ellipses, for
outlined ellipses our calculations are based on magic, and still need to
be pinned down.

The issue is that the outline of an ellipse is *not* an ellipse, and
we must therefore develop a method to calculate the flatness of such an
outline segment, and from this, the maximum delta t on the parametric
form of the ellipse.