calculate last point on offset bezier separately
we calculate the last point on offset bézier curve analytically. because
the subdivision method used for earlier points won't allow us to
calculate the tangent for the last point, we calculate the precise
tangent by evaluating the derivative of the original bezier curve at
parameter t==1.f.

we need to do this so that we don't end up with a gap when adding joints
when connecting the bezier curve with the next curve segment - or when
we add line caps.