tackle special cases for beziers
we must guard against curves where first or last control point is
identical with first or last endpoint. These cases need to be
identified, and tangents need to be calculated by using the remaining
control point, or, if no control points remain, the opposite endpoint.

effecticely, we never want to be in a situation where we want to
normalize a zero-length vector.