rename
le_camera_info → le_camera_settings_info
le_stage_camera_o → le_camera_settings_o

a camera on stage is not a real, interactive camera. instead, it is an
entity which only in combination with its node becomes meaningful. the
same camera can be used with multiple nodes, therefore we are better off
naming this object a le_camera_settings_o, from which we may
*instantiate* cameras by placing a camera inside a node.