refactor stage pipeline generation
+ first create a set of all possible shader define strings,
+ then generate shaders for each of the combinations
+ then attach shaders to pipeline state for each primitive

This makes it much faster to load lots of materials, as most materials
share the same shader and/or shader defines set.

This also means we're only calling the shader compiler on shaders which
are unique, which makes it much faster to load scenes with multiple
materials which use the same shaders.