add descriptor set state cache
We cache DescriptorSet state per renderpass, and we only allocate new
DescriptorSets if we detect that one or more Descriptors should change
from what is currently bound.

This minimises the number of Descriptors we need to allocate because
for buffers we use dynamic descriptors which often reference the same
buffer, and therefore are identical.