upload animation samplers w/ animation_info
animation samplers are owned by animation, therefore the indices are
relative to the current animation, and samplers cannot be re-used for
other animations than the animation they were defined for.

internally, we build a vector of keyframes from samplers which we store
with the corresponding channel, so it's better to keep that information
together in the first place.