add timebase module
time base serves as a source of unique truth about time - we use this to
express time in base units. such a unit, a "Tick" is set at 1/12000 of a
second, meaning, there are 12'000 Ticks in one second.

This allows us to express duration using integral types. We prefer this
over using floats, because with floats we would lose precision at large
values. Using 1/12000 as the base unit means that, if we don't mess this
up, we should get a constant precision for our time measurements.