update timebase api
we want to be able to set fixed interval in `update()` method, if any.
If we call update with fixed interval 0, this means to use the clock to
figure out the interval. If we wanted to have a fixed interval of
actually 0 ticks, this would be the same as not calling update, which is
why overloading the fixed_interval parameter makes sense.