fix a nasty bug
fixes an issue where we were hashing the address of the binding info,
instead of its contents. note the missing & in the correct invocation of
the hash method.

I think this originated from using auto in the loop-header and thinking
that the loop iterator was an object instead of a pointer. We should be
more explicit in loop headers from now on.