[le_stage]: refactor keyframe import
store data for keyframes in array, and additionally store nuber of used
elements of this array. default: 1.

we do this so that we have a common method of dealing with scalar array
(for weights), vec3 (linear blend translation etc.) and vec3 arrays
(cubic spline translation) keyframes.

we calculate the number of array elements per keyframe from the ratio of
outputaccessor.count / inputaccessor.count, and store this with the
keyframe.