[stage]: set attributes using arrays
we use arrays to set attributes which have more than one element - this
is the case for morph target attributes for example. Using arrays allows
us to be more flexible in our shader.

We must, however, correctly calculate the location of each attribute, as
using arrays in locations means that the location(s) directly after the
location used for the array may be used by the array.