[stage]: implement larger joints number
skins may have large number of joints, which are then indexed using the
joints_set in the shader.

In the vertex shader, we use `readonly buffer` (SSBO) instead of regular
uniform buffer objects, because this allows us to not specify the size
of the array of matrices upfront.

When rendering a mesh, we first calculate all the matrices for the skin,
which we then upload. There is room for optimisation here, as a skin is
most likely shared among meshes, (and if it has a skeleton attribute it
can be re-used, as this means that skin calculations are meant to be
relative to the given skeleton node, and not to the node into which the
mesh was instantiated.

If we find a method to optimise this, we could save uploading skin
matrices, and re-calculating skin matrices per-node.