[backend] remove setArgumentData command
+ remove setArgumentData as a full command - this command is a
helper/"macro" which allocates some data from the current linear
allocator buffer, and copies data given to it into the newly allocated
memory. From then on, everything proceeds as if a call to
bind_argument_buffer had occurred, which is what the command internally
expands to.

This simplifies execution, since it reduces the number of code paths.