[backend] fix staging allocator issue
fixes an issue where image and buffer destruction would bypass the
vmaAllocator. Since the vmaAllocator was used to create these resources,
bypassing it means that the allocator then has no way of keeping track
of the resources anymore.

Any resources allocated/created via the vmaAllocator must be
deallocated/destroyed via this allocator, so that it stays in sync.