[backend] add scratchbuffer_sz to blas_info
add field for required scratch buffer size to blas_info - this can be
set when allocating the acceleration structure.

I wish there was a more elegant way to do this, but it appears that in
order to request memory requirements we must already have an
acceleration structure object, which means we cannot set this value
earlier.