[backend] allocate scratch buffer if blas create requested
if rtx acceleration structures need to be built/updated we must provide
a scratch buffer with memory for this purpose - WHY, O, WHY???

Anyway, we don't want to expose this to the user, as we see this as an
implementation detail, and it does complicate the rtx api.

If we're keeping this buffer internally, we can also make sure it is
large enough for all build/update operations scheduled in the current
frame - and we can also make sure that it is properly synchronised
internally.