expose rtx_shader_group_info type
this type is used to structure shader group information for rtx shaders.
we also use it as part of the hash calculation over rtx pipelines.

this was previously declared in backend, but we must declare in
renderer_types, so that user may specify these parameters when creating
an rtx pipeline.