[backend] require le_device_o to initialize shader_manager
shader_manager requires a le_device_o (which it will co-own via
reference-counting). This is so that within the shader_manager we have
access to properties queried from the physical device associated with
the device.

We need to know the value of these properties for RTX so that we can
calculate the memory requirements for shader handles, and shader binding
tables, for example.