[encoder] implement construction of sbt
we construct the sbt for ray_gen, miss, hit, and callable shaders by
writing out shader instances (shader + associated parameters, if any)
into GPU backed scratch buffer.

We keep track of byte offsets into scratch buffer when doing so, which
allows us to use the scratch buffer (via offsets) as the buffer used for
shader binding tables.

Note that this means that shader binding tables are ephemeral, and must
be created per-renderpass, and per-frame.

I think that should be okay for now, and we can think of methods for
caching shader binding tables by, for example, copying the buffer slice
used for storing the sbt to a dedicated GPU-only buffer.