fix shader binding table generation
when uploading the shader binding table to the scratch buffer, we were
not moving our data pointer to the correct position/offset. Since the
blob with the shader handles contains a header we must skip the header
in any calculations which copy shader handles.

we do this by defining shader_group_data_payload, which defines the base
offset for shader group handles in the blob. any acceess to shader
handles from the blob needs to be relative to this address.