fix rtx pipeline recompilation issue
RTX pipelines are compiled to vk pipelines in record stage. If pipeline
gets recompiled, the currently executing backend process stage will most
likely (race condition) get the latest version of the pipeline when
querying based on latest rtx_pso hash.

This will lead to device loss, because the shader binding table for the
backend process frame will have been built based on the old pipeline,
and not the (new) pipeline which will be bound now.

We fix this by recording the vulkan object in the command stream,
alongside the known pipeline layout state. We know all this as a product
of compiling the pipeline.

This allows us to keep the rtx pipeline synchronised with its rtx shader
binding table. Ideally, we could do this with graphics pipelines, too,
but graphics pipelines have the horrible property that the renderpass,
and subpass id is part of the pipeline state, and we can therefore not
compile the pipeline in advance...