[rtx] port ray_tracing_nv → khr_ray_tracing
Update backend to use khr_ray_tracing instead of Nvidia's
ray_tracing_nv. This means that we had to require Vulkan 1.2, and some
features which are dependencies of khr_ray_tracing, such as the method
to get device addresses for buffers, which is used by khr_ray_tracing,
instead of nvidia's opaque handles, and plain VkBuffers.

We had to modernise the way extensions and features are requested, which
is to be found in le_device.

Instead of NV raytracing methods we're now importing khr_raytracing
methods via the vulkan beta headers. We should probably put this behind
a header guard, and make it conditional on LE_FEATURE_RTX