[rtx] update vma
We need the latest version of the vulkan memory allocator, as we need to
be able to allocate buffers with
`VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT`. This is a requirement for
some buffers, notably the scratch buffer, vertex buffer, index buffer,
and the buffer containing the data for the shader binding table.