[renderer|backend] add opaque le_texture_handle
Instead of being treated as resources, we give textures their own
dedicated type, an opaque handle.

The handle is a pointer to some optional debug info (we decided against
using constexpr hashed strings because the debug experience is better if we make the
opaque handle point to debug information). There's also no risk anymore
of hash collisions, and it feels better to pass handles around if more
than one module wants to reference the same texture.

Since texture handles are unique based on the string they are
initialised with, it is still possible to refer to the same texture
without passing the handle through, but by independently requesting a
handle for the same identifying string.