[renderer] change texture usage signature+impl
change texture usage so that providing the texture handle is enough when
setting up a renderpass.

when a texture is declared, you must provide an image_sampler_info. This
image_sampler_info is then copied into the renderer, and associated with
the texture_handle.

You must specify an image handle which your texture refers to with the
image_sampler_info.