Revert "[renderer] change texture usage signature+impl"
This reverts commit bfcd11b0cdf8b5ee48177e26658703dfd36fe08e.

We're reverting this, as storing sampler/view information with a texture
handle is a bad idea, as it creates a source of contention: many callers
or renderpassed might want to have different sampler/view information
associated with a texture handle, and this is impossible to do cleanly
if there is a global store sampler/view information.

we want the link between texture handle and samlpler/view info to be
ephemeral, and we want it to be pass-local (owned by pass, destroyed
with the pass with the frame).