[le_path] update signature get_[vertices|tangents]_for_polyline
Update syntax so that caller is responsible for allocating object into
which to write result. We don't want to hand out pointers from inside
the path module for whose lifetime there is no guarantee.

As an additional benefit, this allows callers which call the same
function multiple times to re-use their scratch memory, thus reducing
allocations.