[backend] limit auto-creation of per-frame image views
Only auto-creates ImageViews for images which are used ( SAMPLED |
STORAGE ) if these images are used by a compute pass. Images used in a
draw pass are usually introduced using useTexture(), which we handle
later in this method, and for which we have much better information at
hand.

We should only create the ImageViews which we actually need, and
therefore it's better not to auto-create ImageViews, just in case. On
top of that, some imageViews will be missing information at this point,
information such as the ImageViewType.