[core] rewrite callback forwarding
re-implement callback forwarding using a more explicit low-level
approach. By writing machine code ourselves, we become independent of
the inline assembler. This allows us to clean up the code a bit.

We're making callback forwarding dependent on the target architecture
being x86_64, which is the architecture for which we have written the
assembly code.