[le_renderer] add readback to set_[vertex|index]
Add an option to capture transient binding state for vertices and
indices set directly from within encoder - this should make it possible
to re-use dynamic vertex and index data, as long as it is used with the
same encoder / or even the same frame.

I think this can be useful for 2d drawing for example, where we might
end up with batches of primitives which have the same geometry, but are
not immediately drawn following each other. In this case, we would have
had to generate geometry for changing primitive. With this readback
option, we can now implement something like a cache, where we generate
primitive geometry only once per frame per primitive.