[multi-swap] window to swap, render_complete
* only khr swapchains use windows, it makes more sense to make window a
property of khr swapchain than to let backend have special reference to
"the" window, as there is no special window, there should only be
swapchains, and render targets for what the backend is concerned

* we need a render_complete per swapchain - as semaphores can only have
a single "waiter" - waiting on a semaphore puts it back into the
unsignalled state, therefore no more than one single client can wait on
a semaphore. We therefore give each swapchain a separate render_complete
semaphore.

On render queue submit we say that we want all `render_complete`
semaphores for all swapchains to be signalled on completion. We could
add more granularity here (submit in smaller batches such as
per-renderpass) and signal the semaphore as soon as complete - but let's
keep things simpler for now...