[rendergraph] infer pass extents from swapchain
if a pass has no extents given, we try to infer pass size from any image
attachments which are swapchain images (because these have a known
size).

If a pass does not map any swapchain images, and has no extents given,
the extents of the first swapchain image are taken.