[shader_compiler] fix error message parsing
use regex instead of sscanf to parse error message. Windows error
messages may contain colons (':') as part of their paths, and cannot be
parsed in the same simple way as linux error messages, where colons
cannot be part of the path and can therefore be taken as delimiters.

I hope std:regex does not add a ton of compilation time and build
complexity on linux - otherwise we will have to hand-parse the error
string.