[win64] update file watcher
filter file modification events, and cache events, so that we trigger
fewer callbacks.

some file system events are reported in great granularity by
ReadDirectoryChangesW, and we don't want to trigger a callback for each
of the small changes affecting a file. Instead, we opt to only trigger
on the last event concerning a file.

This is achieved by internally filtering file change events.