[win64-hotreloading] implement for win64
implements hot-reloading for win64 targets - this will work even if a
debugger is attached to the app.

there is a vexing bug in visual studio which made this quite a
rigmarole: once a library is loaded into the debugger, the debugger
won't relinquish the library's .pdb file - even if the library gets
unloaded.

We get around this by first seeking for any processes which have a
handle to a stale .pdb, then forcibly taking ownership of the handle,
dropping the handle, and then, finally, deleting the stale pdb once
there is no more handle to it.

The technique for this is based on what Stefan Reinalter described in
his blog here: <https://blog.molecular-matters.com/2017/05/09/deleting-pdb-files-locked-by-visual-studio/>

It's a wild hack, but it works \o/

It won't affect production systems, since hot-reloading is by definition
only used in a debug context.