[win64] add powershell script for auto-recompile
add a powershell script which watches a folder for file changes, and if
changes are detected, triggers ninja in a build folder. This script
makes it possible to trigger recompiles automatically whenever a source
file is saved.

This, together with hot-reloading switched on, can make for a nice
development environment, where each file save triggers a recompile,
which in turn triggers a hot-reload.