[ecs] add level of indirection for entities
keep entity id to be increasing monotonically, but decouple entity id
from index of entity inside of vector.

this makes it possible to remove an entity even if it is not the last
element in the entity vector, which is something a user of the API would
have assumed to work in the first place.