[ecs] change signature for entity_component_at
entity component accessor acts in a similar vein as the [] accessor for
a map, viz. if accessing a not yet present component type, the missing
type is added to the entity.

we probably need a separate entity_has_component method so that we can
test whether an entity has a specific component without modifying the
entity.