[backend] fix missing swapchain extensions bug
when multiple swapchains requested device extensions, only the
extensions requested by the first swapchain in the list would be added
to the list of requested device extensions.

This lead to crashes if, for example, the first swapchain was an image
swapchain and the second swapchain a khr (window) swapchain. In this
case, the image swapchain would request no device extensions, and the
window swapchain would request `VK_KHR_swapchain`, but that would be
(wrongly) ignored. Calling vkCreateSwapchainKHR subsequently would lead
to a SEGFAULT because the call would not resolve because the extension
was not requested.

We're fixing this by going over *all* not just the first swapchain when
querying for requested device extensions. We're already doing this for
instance extensions.