[le_pixels] add validity check
allow le::Pixels to test itself for validity - this is necessary so that
we can detect whether a le::Pixels object was successfully created.

Reasons why le::Pixels objects might not have been created successfully
might be that the image file was not found, or the file was no image.