[le_core, windows] fix build for windows/debug
windows needs le_file_watcher to be dynamic, otherwise all sorts of
strange things happen. we can work with this - instead of loading the
filewatcher as a static module we're loading it using the default
setting - dynamic when in debug, static when static. This also means
that we must query the api interface by using the dynamic interface in
le_core.