[api_loader] auto-detect logger presence
print via logger if logger was loaded - otherwise, print via fallback
method. a logger may not be loaded because the logger module might be in
the middle of hot-reloading.

Since before hot-reloading we're now wiping the api interface struct, we
can test for whether the api interface struct for le_log is set, if not,
we must assume that the logger module has just been unloaded.