[le_core] add global key-pointer store
adds a globally accessible key-pointer store which allows us to store
key-retrievable pointers across api interface reloads. I'm not totally
sure whether this is the right API, but it is useful so far.

Currently, we need this so to test whether the backend has been
reloaded, in which case we must re-patch external proc addresses (see
le_instance_vk:patchExtProcAddrs)