[file_watcher_linux] fix remove_watch
inotify may return identical watch_id for multiple files, if files live
in the same directory. in Island, we're watching files individually -
and we want to be able to remove individual watches.

previous to this fix we were removing watches by removing the inotify
watch associated with the file - without checking whether the inotify
watch was shared with other files... hilarity ensued.

This fix decreases the number of files attached to a inotify watch, and
if zero, removes the inotify watch, which is the correct behaviour.