[file_watcher_linux] protect against deadlock
implement a command-buffer style pattern to protect against possible
deadlocks in case watches are either removed or added from within a
watch callback function.

any request to add or remove which cannot be immediately processed
because the lock is held is added to a list of deferred commands. these
are processed at the end of the polling loop, once the lock held
throughout polling has been returned.