[backend] implement spirv_reflect codepath
use spirv_reflect for shader code reflection. We are only using a
fraction of what SPIRV-Cross offers, and we're only using it for
spirv-code reflection.

spirv-reflect appears to be a more fitting option for this particular
use-case, especially since it's lightweight.

This commit allows you to switch between spirv-cross and spirv_reflect
via #defines.

We're going to do some more testing and will eventually deprecate
spirv-cross. This will also be beneficial to compile times.