[backend] store shader stages with pipeline info
store shader stages which are active for a given pipeline layout with
the pipeline layout info. we do this so that we can keep track of this
state during command buffer processing.

pushConstant command needs to know for which shader stages given push
constant data shall apply. By default, we activate all shader stages
which are in use with the currently bound pipeline. For this we need to
know which shader stages are active for a given pipeline.