[shader_compiler] make error parsing more robust
allow for missing source file in error message. Because of an internal
glslang bug error messages which concern constructor methods are missing
file path and line information. Therefore we cannot assume that these are
always present.

I repoted the bug, and sent a PR to glslang, hopefully the fix will get
upstreamed soon.